Seperate the optimization tests from the generic propeller tests
